Fix yank command argument bug
Invoking `yank echo' would be treated as a valid invocation causing
`echo' to be used as the yank command even though the `--' is not
present. This is not compliant with the documented and intended
behavior. Therefor ensure presence of the `--' option.

Ensure presence of the `--'
option when a yank command is supplied.